Safari-611.1.10/LayoutTests/fast/canvas/canvas-shadow-expected.txt

Some tests for Canvas shadows

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[4] is 0
PASS imgdata[5] is 128
PASS imgdata[6] is 0
PASS imgdata[4] is 0
PASS imgdata[5] is 128
PASS imgdata[6] is 0
PASS successfullyParsed is true

TEST COMPLETE
